move componenets to their own folder
