--- src/Components/Plots/Plot1.tsx
@@ -1,9 +1,0 @@
-function Plot1() {
-    return (
-        <div>
-            <h1>Plot1</h1>
-        </div>
-    );
-}
-
-export default Plot1;
